kaveri asopa  1st pr
Getting Started as a Front-End Developer
We've already talked about what the front end is and what a front-end developer does. Let's now examine some of the key qualifications and abilities needed to qualify as a front-end developer.

It's crucial to realize that while you don't need to be an expert in everything to become a front-end developer, you do need to be familiar with the fundamentals, such as HTML, CSS, and JavaScript.